Update patch set 3
Patch Set 3: Code-Review-1

+1 for tox; -1 for pbr.

I think that adding jenkins validation of pep8/pyflakes (and soon some unit tests) is a great idea, but making the project installable is putting the cart before the horse. This is going to be an internal WMF tool and we aren't likely to be installing it via pip or even a deb based off the egg.

Patch-set: 3
Label: Code-Review=-1