Update patch set 3
Patch Set 3:

Hashar wrote:
> If we want to run unit tests in a tox maintained virtual env

I didn't see any obvious signs in Jenkins output that we are currently using virtualenv in the test runs. I can see the usefulness of virtualenv in testing if we start adding dependencies on external python libraries, but until we reach that point it seems unnecessary.

It actually seems that we are getting testing using pep8 & pyflakes independent of this patch. Can we make a simpler change just for the E128 configuration or is the tox testing setup the preferred way to test python code now in Jenkins? Would it be enough to add a setup.cfg file with a pep8 configuration section?

Patch-set: 3