Rearrange modules to make more sense
Get rid of the scap module with the contents being moved to tasks and
main. Main is the module for command line wrapper functions. Tasks is
the module for business logic. Utils is the module for general purpose
utility functions (array builders, generic commands, etc). Using this
organization the check_syntax function was moved from utils to tasks and
renamed check_php_syntax to clarify it's purpose.

Change-Id: Ibd869dab99de03b9a983ee38269f9445d437eee8