Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

Very happy to see this.

For later: perhaps there's a case to be made for a task context manager that emits a log message on entry, and on exit emits the duration and logs it via statsd.

   with task('sync_common') as t:
        t.targets = 'mw*'
        task.run()

(I'm not convinced it's the right API, just messing around.)

Patch-set: 2
Label: Code-Review=-1