Update patch set 3
Patch Set 3: Code-Review-1

AFAIK we retain the ability to run the "old" scap by running scap-old, which is a nice safety to have. I share the enthusiasm for killing scap-2, but if keeping it around allows us to have a bit more safety, why don't we?

Also, I don't see the logic in replacing find-nearest-rsync with code that shells out to ping. Replacing a PHP implementation with a unified Python implementation is one thing, but here you're just shelling out to another executable, and one that is lest optimally suited for the task.

Patch-set: 3
Label: Code-Review=-1