Update patch set 3
Patch Set 3:

Keeping scap-2 to retain old-scap functionality is fine. I'd already forgotten that we'd done that.

My reason for killing find-nearest-rsync is to get rid of a `sudo -u root` call to a random script that could be easily changed. Parsing /bin/ping output is more fragile, but less scary to me than running a perl script as root.

Patch-set: 3
Label: Verified=0