Add cluster_monitor ssh wrapper
Add scap.ssh.cluster_monitor function to run a command on the cluster
via ssh and monitor progress with a live updating progress message and
real-time error logging. Progress messages are sent to stderr. Failed
commands will be logged as warnings as they occur. Returns the count of
successful and failed commands.

Bug: 59922
Change-Id: I0ddefc136ce3d136599571288e8ab056a5a889ef