Update patch set 7
Restored

Ori was right and I was wrong about the use of --versions. My desire to make scap only sync the active versions by default relies on each destination server being up to date with the original server modulo the changes that are being staged for all servers. This base assumption fails when a new server is added to the pool or a server rejoins the pool after having been out of service for any length of time. Only sending the active versions of MW means that can't be trusted to be in sync for serving bits.wm.o content.

Patch-set: 7
Status: new