Replace cluster_monitor with a ssh.Job class
Convert simplistic ssh.cluster_monitor function into the ssh.Job class
which allows running a remote command to be progressively described
using syntax that makes the intent of the final operation more clear to
readers.

The new log.ProgressReporter class can be used to display progress
messages for cluster_ssh or other multi-step tasks that have a concept
of success, failure and steps to be performed.

Change-Id: I5580e3a520aac8126b569188844c7aa5db3a93d6