Update patch set 1
Patch Set 1: Code-Review-1

Something is wrong when you are parsing state out of a string that was generated by another part of your program. Shell command lines are not a data-interchange format. You should break this up and be disciplined about not coupling unrelated things in this fashion. It looks like you need a progress indicator class (it is important that it know nothing of shells, hosts, etc.). It also looks like you need representation of a task that is richer than a command-line string or list.

The functional style that you've adopted, which keeps state in closures rather than objects, works well when you are disciplined about keeping functions very small and stingy with side-effects. But the approach is beginning to seriously fray. IMO, you need to either aggressively refactor this function (and several others) into smaller pieces that have discrete responsibilities, and you should probably bite the bullet and have a Task class. The benefit of having objects is that you don't have to know everything right away, as you do when you are calling a function. You can build up the object gradually (setting a description, configuring logging and error handling behavior, etc.) by getting and setting properties, and defer compilation to a command-line until you're really ready to execute the task.

Patch-set: 1
Label: Code-Review=-1