Use common MW_DSH_ARGS in sync-wikiversions
sync-wikiversions seems to be the only sync script that doesn't use the
common dsh arguments. Not coincidentally it is the only script needed
for the weekly new branch deployment that failed for me on 2014-02-27.
The failure was caused by the full fanout dsh overwhelming my ssh-agent.

The difference between the prior dsh arguments and the common settings
from mw-deployment-vars.sh is the addition of `-F30` to limit the fanout
to 30 concurrent hosts and raising the ssh SetupTimeout option from 10
to 30.

Change-Id: I86ec2f36ac9a0f5e6f6aa34c26d5fd4676acd84c