Update patch set 2
Patch Set 2:

I don't know what Jenkins' problem with this is. I tried checking out the prior patch in the chain (Ie50be09) and cherry-picking this patch on to it and gerrit rejected the push with the message "No changes between prior commit 0ea3a36 and new commit ea79b50".

Patch-set: 2
Label: Verified=0