Delete stub l10n file after generating ExtensionMessages
The bootstrapping process added in I5467ac8 generates an en l10n file
before the ExtensionMessages index file is created. This file needs to
be deleted before rebuildLocalisationCache.php is run or it is likely to
be seen as being up to date based on the file's timestamp. This leads to
missing l10n values for extensions.

Bug: 51174
Change-Id: I3533aacea048c80ab04bdd83d83dff98acfc1ca2