Update patch set 3
Patch Set 3:

> [...snip...] use the command basename as a name for the ProgressReporter

I did something like this in patch set 1. Ori had the opposite reaction and wanted the concerns of reporting and execution. Actually I think Ori is looking for an even higher level of abstraction, but I was hoping that this would be seen as a small step in the right direction.

I think ultimately this is headed towards an abstraction for executing a unit of work across a group of hosts that can be specialized for different command transports (ssh, salt, whatever) and different reporting needs. I'm interested in breaking the monolithic task.scap method down into smaller units of reusable business logic and I hope that while doing that better abstractions become apparent.

Patch-set: 3