Update patch set 1
Patch Set 1:

> I don't know what "NOQA" means. I checked <https://en.wiktionary.org/wiki/NOQA> and <http://www.urbandictionary.com/define.php?term=NOQA>.

In this context, "NOQA" was used as a hint to the flake8 python source code checker [0] asking it to ignore a warning triggered by the line containing the comment [1]. The particular warning intended to be ignored is the  flake8 warning that the imported functions are unused in the module.

> I'm also not sure what any() does, but I imagine I could probably find that answer out on python.org somewhere.

`any()` is a python built-in function that can be used to find out of any item in an iterable is true [2]. It is being used here as a low side effect method for "using" the imports so that flake8 will not issue an unused import warning.

[0]: https://pypi.python.org/pypi/flake8
[1]: http://stackoverflow.com/a/10506715
[2]: http://docs.python.org/2.7/library/functions.html#any

Patch-set: 1