Update patch set 3
Patch Set 3:

I have working code for a different take on this that I think moves further in the direction that Ori was thinking about with his comments on patch set 1:

            # Create a job for a list of hosts
            update_apaches = ssh.Job(mw_install_hosts)

            # Remove some hosts from the list
            update_apaches.exclude_hosts(scap_proxies)

            # Shuffle the host list
            update_apaches.shuffle()

            # Set the command to execute
            update_apaches.command(
                ['/usr/local/bin/sync-common'] + scap_proxies)

            # Enable a progress report while the task is executing
            update_apaches.progress('sync-common')

            # Run the task
            update_apaches.run()

All of the methods of ssh.Job are chainable (return self) as well. I'm not sure that pep8 really likes chained method calls in the same way that the java world does though.

Patch-set: 3
Label: Verified=0