Add script to cleanup l10n cache for an inactive MediaWiki version
This is a new maintenance script that can be used during the weekly
"train" deploy cycle to reclaim disk space used for l10n cache in old
branches. In the 1.23wmf17 version checkout the full l10n cache (CDB and
JSON files) consumes 1.6G of disk space. This cache is only needed when
the branch is being actively used to power a wiki. The branch will
remain on the deploy server and cluster hosts for at least 31 days after
the wikis have moved on to a newer version to satisfy static asset
requests from the front end cache servers.

Change-Id: I5f9299bd77278b79aeb288be97638b39aea87b4f