scap-recompile: always point to most recent version
The scap-recompile script is bugged in beta.  It get the 2nd version of
MediaWiki in use to forge the destination path.  On beta there is only
one version, and the 2nd version ends up being undefined.

The scap-recompile provided by wikimedia-task-appserver used to work. It
got removed from that package...
https://gerrit.wikimedia.org/r/#/c/109950/

...and added back to puppet THOUGH IT GOT ALTERED!
https://gerrit.wikimedia.org/r/#/c/109951/

And broken for beta.

This patch use bash trickery to get the last element of the mwversion
array instead of the 2nd one.  That fix beta, I did test it on
deployment-bastion.pmtpa.wmflabs.

Change-Id: I595ea187f85ea3713c4acdf6c8a258945fcf564e