Compile wikiversions.json to cdb for local sync
MWMultiVersion reads the compiled wikiversions.cdb file rather than the
wikiversions.json source file. PHP operations that happen on the deployment
host during an execution of scap need to have access to a compiled
wikiversions.cdb in the local deploy directory (eg
/usr/local/apache/common-local) that matches the versions set in
wikiversions.json from the stage directory (eg /a/common) and the
deployer's expectations. Add a compile step for wikiversions.json before
calling sync-common locally and make sure to clean up the staged
wikiversions.cdb file prior to syncing with the rest of the cluster.

Bug: 63659
Change-Id: I9ac29bdaef240d5c55dd1ee38113dd4a0d4469e3