Make logging destinations configurable
Change logging system setup so that the settings file and/or command
line options can control enabling/disabling and destination hosts for
udp2log and tcpircbot appenders. This will allow the beta cluster to
send it udp2log output to the appropriate location as well as disabling
the tcpircbot there as it's unlikely that any channel wants to be
spammed with the nearly continuous rebuilds triggered by Jenkins.

Bug: 63746
Change-Id: I557a0e4770b0be4f0e3971093c296cd54920c9bd