Update patch set 1
Patch Set 1:

> Whenever someone copy paste the output of a scap run, the coloring information is lost :)

That's a good point too. Personally I think in a cut-n-paste scenario I would be ok with not having the colors or log level names. The log level names could always be retrieved from the json log message on fluorine and in logstash if needed for some reason.

Patch-set: 1
Label: Verified=0