Exit with non-zero status on soft failure
By design scap is tolerant of the failure of a number of hosts to sync
with the deployment master server. Typically a human will be looking at
the failure status in real time during a production deploy and can
evaluate if the failures are too severe. When used in the beta
environment, scap is invoked by Jenkins and thus needs to alert a human
to analyze the results in case of a partial failure. Exiting with
a non-zero status will allow Jenkins to detect a potential problem with
a given run and raise an alarm.

Change-Id: I9d92d08a2c75bc1e3950196a1a91cabcc46b778b