Make `--verbose` a universal argument
Change the logging usage in scap so that all "useful for deployer to
see" log messages are emitted at info or a higher level. Raise the
default logging level for the console appender to INFO. Make the
`--verbose` command line argument universal in cli.Application.

This change also removes the short form `-v` argument for the
scap-purge-l10n-cache script as it conflicts with the short form `-v`
verbose argument.

Change-Id: Ifcfcbfe1848d4b297faa6c712c8b21f15e0f92a7