Update patch set 1
Patch Set 1:

> Maybe you're not creating the object at the right time, then?

The setup dance is a bit complicated at this point. I think it might be something that could be addressed along with the cli redesign proposed in bug 65827. Currently loading the config file is deferred until the cli arguments have been parsed to allow specifying an alternate config file and/or individual config overrides. Processing the cli arguments in the current design is dependent on constructing the scap.cli.Application object so that the Application can add optional and required arguments and post-process the parsed arguments.

Reorganizing the cli as a single command with sub-commands would allow hoisting the shared arguments to the core command where they could be processed and used to bootstrap the application's internal environment prior to constructing the sub-command.

I actually like the way that I have used the lazy properties that this patch is removing, so I'm not strongly motivated to embark on a major surgery to the initialization order of the current class hierarchy just to land this particular patch.

Patch-set: 1
Reviewer: BryanDavis <1001@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0