Partially revert "Convert sync-dir and sync-file to python"
This partially reverts commit 28a4f0efa2 / change I5ab7b2fba, restoring
sync-dir, sync-file, and sync-common-file to the shell script versions.
The new version didn't work for Sean Pringle when he attempted to sync a change
to the database manifest (he was able to use these old scripts). I wasn't able
to reproduce the issue, but it's safest to simply revert this for now until
Bryan can look at this in detail.

The docs/scripts.rst changes are required to make Jenkins happy; otherwise it
tries to invoke "sync-dir/sync-file --help" so that it can interpolate the
output into the generated documentation.

Change-Id: I449eba1bc6ea1783d2978c58ae75d7cf592f94ad