Update patch set 1
Abandoned

I have applied this in beta via a cherry-pick but it is not needed or wanted in production. Ideally we will end up getting server certificate collection working in beta at some point so we can drop the cherry-pick there too, but I'm not holding my breath for someone to find time to fix that.

Patch-set: 1
Status: abandoned