Use -n flag on all sudo calls
"The -n (non-interactive) option prevents sudo from prompting the user
for a password.  If a password is required for the command to run, sudo
will display an error messages and exit." -- sudo man page

We were using this in some places already but not consistently. The fail
fast with an error message aspect of this seems better than the
possibility of hanging with a password prompt in the rare case of a user
who has missing sudoer rights.

Change-Id: I39d9f61dbae9c13f4944a73d649ef7ebb9d3f074