Update patch set 2
Patch Set 2:

(1 comment)

I haven't run it, but it looks good superficially. I think config should be added to use these options and the keyholder service in beta so we can fully test. Cherry-picking is a trivial way to fully test a change like this in beta before merging.

Patch-set: 2
Label: Verified=0