Update patch set 1
Patch Set 1: Code-Review-1

In the face of ambiguity, refuse the temptation to guess. If SSH_AUTH_SOCK is set in the environment, it should be used. If it doesn't work, we should fail loudly. I think bd808's suggestion to make the socket usable for l10nupdate is better.

Patch-set: 1
Label: Code-Review=-1