Update patch set 1
Patch Set 1:

The change in I5782e8c which forced replacing the existing SSH_AUTH_SOCK setting from the env is the root of this fix. If instead the shared ssh agent was only used if SSH_AUTH_SOCK wasn't set the bug would be fixed too.

Patch-set: 1
Label: Verified=0