Make scap localization cache build $TMPDIR aware
Introduces a contextmanager that creates a temporary directory and
cleans it up on block exit. The directory is made under the system temp
dir and attempts are made to ensure that the created directory is
unique. The contextmanager includes try/finally wrapping that should be
more effective at cleaning up the temporary directory when scap is
interrupted than the prior implementation.

Bug: T97257
Co-Author: Bryan Davis <bd808@wikimedia.org>
Change-Id: I1d8e4eddc308260f8e2fd488cf84dcf82da3734a