Better handling for php lint checks
* Only lint check sync-file targets that have php extensions
* Allow php files to start with doctype declarations too

Before this change, sync-file for wmf-config/interwiki.cdb would fail
because it was checked with `php -l` (which it fails badly
unsurprisingly). There are also several *.php files in docroots and
other locations that would fail a sync-file push because they start with
a '<!DOCTYPE' and other HTML content before using PHP blocks to insert
dynamic content.

Change-Id: I36e495d0365480fd1dce2e2485135f0a4273bf7a