Add service deploy via scap
Deploy a service that is currently deployed using trebuchet with scap.

Should now be able to run the `deploy` command from
`/srv/deployment/{repo}`. Reads the `scap.cfg` file from the cwd
(provided the cwd is under `srv/deployment` and is a git repo)

The `scap.cfg` file should contain the following information (scoped
appropriately for the project):

    [global]
    git_repo: restbase/deploy
    dsh_targets: restbase-installation

    [wmflabs]
    git_server: staging-test-tin.staging.eqiad.wmflabs

The deploy command:

1. Tags the current revision in the current git directory
2. Runs `git update-server-info` for the current git directory (as well
   as for any submodules)
3. SSHs into each server in the `dsh_targets` file
4. Runs git fetch in the `/srv/deployment/{repo}` directory of each target
   (running checkout if it does not exist)
5. Checks out the tag created in step 1 on each of the target machines
6. If a `service_name` is specified, the service is restarted
7. If a `service_port` is specified, make sure that it is accepting
   connections

Change-Id: I542bb9acc0f65360452fb4f0b0c92dec4894e5d8