Update patch set 5
Patch Set 5:

(4 comments)

The only problem with just using the context manager without the global function is that you wouldn't be able to get a logger without creating a new context. That's probably fine though, and it would make things simpler.

Also, maybe this should live in utils since the other context managers are in utils.py and it would solve the relative import problem. But then we'd have logging related stuff in utils instead of log.py

Patch-set: 5
Label: Verified=0