Update patch set 7
Patch Set 7: -Code-Review

Would be a nice to have to just put the final finishing symlink in-place with a config var like `final_path`  (ln -s [deploy_dir]/current -> final_path)

That way we're not dependent on somthing/someone external to scap to manage symlinks. The reasoning there would be to ease transition between trebuchet and scap.

Removing -1, leaving it up to you.

Patch-set: 7