Update patch set 1
Patch Set 1: Code-Review-1

Thanks Chad. Looks like I am owed a slap for this :) However, the third change (If3bd73f6b4) should not be reverted -- that bit was successful. We are using wikiversions.php in production. It's the wikiversions.cdb code that ought to be (carefully) removed.

Patch-set: 1
Label: Code-Review=-1