Filter target host logging from stdout of main process
Change I60e1eb4b630d4779e2659310475b1973c783c6d8 implemented structured
logging from the target host to the deploy host which resulted in a lot
of noise on stdout of the main process. The handler for stdout has been
configured with a new filter to omit everything from the targets. We
should work on implementing specific panes or log files for such output.

The generic `log.Filter` class should help when it comes time to
implement such filtering of other handlers (especially if we're to
accept user provided filters) as it implements a very general syntax. Or
it's just crazy and unnecessary. Who knows? I'm very tired.

Bug: T113779
Change-Id: I2931b024fdb1f80fc4a84f8550d5a1658da426c3