Implement basic idempotence
Summary:
Adds a `self.noop` flag that is set when it can be determined that a
stage no longer needs to continue to execute. Uses info-level logging to
inform deployer (Soon™).

For `config_deploy` the fetching and rendering of all templates will
still need to happen on each target to determine whether there has been
a change in the rendered config. This is because templates are combined
with information on the targets before they are rendered. For all other
stages (`fetch`, `promote`) a noop determination can be made much more
quickly and on the basis of the existence directories on the target.

Also adds the `-f|--force` flag to `Deploy` and `DeployLocal` to allow
override of idempotence.

Change-Id: I2f5df09cfd7e44dcf09cfb56f414314c72b12f8b
Bug: T109513

Test Plan:
Run tox
run deploy-local on targets and watch info output

Reviewers: demon, mobrovac, fgiunchedi, mmodell, dduvall, #release-engineering-team

Reviewed By: mmodell, dduvall, #release-engineering-team

Differential Revision: https://phabricator.wikimedia.org/D17