Prevent deployment when a sync.flag exists
Summary:
sync.flag is a persistent lock file that prevents scap from
syncing to the cluster. The sync.flag is created by a deployer
when the staging directory is in a bad state.

When a deployer is about to make modifications to the staging
directory they can create a file named sync.flag to alert other
deployers that it is not currently safe to sync to the cluster.
Scap will enforce the sync.flag by mentioning it's existence as
well as the username of the file's owner. The file must be removed
before any deployments can continue.

Test Plan: `touch /srv/mediawiki-staging/sync.flag && scap`

Reviewers: #release-engineering-team, bd808, Krenair, thcipriani, dduvall, greg

Differential Revision: https://phabricator.wikimedia.org/D36