Rename non-symlink directory to ISO8601 dir
Summary:
Instead of overwrite a current directory that is in the final checkout
location on targets, we rename that directory. There is a problem,
currently, if this renaming must happen multiple times, in that case
`os.rename` refuses to rename the directory to directory.trebuchet
because that directory already exists. This patch, instead, will move an
existing directory to directory.ISO8601_date.

Bug: T119202
Change-Id: I17cabb4b7173b5b60489f54f4549ad228e5a9263

Test Plan:
1. Remove final repo checkout on a target
2. Make a new empty directory in its place
3. subsequent deploys should move that directory

Reviewers: mmodell, dduvall, mobrovac, demon, #release-engineering-team

Reviewed By: demon, #release-engineering-team

Maniphest Tasks: T119202

Differential Revision: https://phabricator.wikimedia.org/D56