Begin to remove `git_deploy_dir` assumptions
Summary:
We don't want to enforce a directory structure where
`config['git_deploy_dir']` has to be the root of all deploy-type
applications. For one thing, since `/srv/deployment` on beta is actually
`/mnt/srv/deployment` the `deploy` command blows up. We want to enforce
`git_deploy_dir` on targets, but not on hosts.

This attempts to remove that assumption for both `deploy-log` and `deploy`
while still keeping it as a strict requirement for `deploy-local`.

In the docs the items, `git_deploy_dir` and `git_repo` were updated. I
think that this giant table, while it looks nice in its final state, is
going to create some painful diffs every time we touch it.

Bug: T116207
Change-Id: Ib209cc87e1f2a7ffbdbf67ec985975b99ea7c45c

Test Plan:
1. Make sure that normal deploy works
2. Try deploy from deployment-bastion

Reviewers: mmodell, dduvall, demon, #release-engineering-team, mobrovac

Reviewed By: mobrovac

Maniphest Tasks: T116207

Differential Revision: https://phabricator.wikimedia.org/D58