Add security-patch-check script
Maniphest Tasks: T118477

Summary:
This is an additional script for now. It may be added as part of full
scap if the method proves to be a reliable way to detect un-merged
security patches.

This does rely on the creation of a version directory for
`/srv/patches/{core,extensions}` i.e.,
`/srv/patches/1.27.0-wmf.9/{core,extensions}` This directory could be
created as part of `checkoutMediaWiki`, for now it can be manually
created.

This patch will collect all the patches for mediawiki-core and each
submodule in reverse sort order, then it will run: `git apply
--check --reverse`, using the collected diffs as the stdin of that
command. In this way, one patch that depends on another patch should
reverse-apply cleanly. If any reversal does not apply cleanly then
patches have not been applied.

Fixes T118477

Change-Id: Ife16245d66d420540d3423077155b25fd7d9795e

Reviewers: demon, csteipp, dduvall, mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D80