Fix submodule rewrites
Summary:
The current method of submodule rewrites assumes that path and name will
always be the same. In some instances that will not be the case. This is
a rewrite of the function using the output of
`git config --list --file .gitmodules` to determine the name and path to
a submodule.

Change-Id: I182c1e4015f8d86c72c9eeffa7882524564a1eef

Test Plan:
1. Add a submodule using: git submodule add --name blah path/other/than/blah
2. Verify that targets fetch new submodule from git_server

Reviewers: dduvall, demon, #release-engineering-team, mmodell

Reviewed By: #release-engineering-team, mmodell

Subscribers: jenkins, Ottomata

Differential Revision: https://phabricator.wikimedia.org/D82