Get rid of sync-docroot
Summary:
It's kind of silly that we have script that basically calls sync-dir
without any sort of sanity checking.

It skips linting(!!), which is kinda ok because it's nothing but
symlinks here...but imagine if someone checked real code into this
path...it'd be Chad + Brion deployments from 2008 all over again.

Reviewers: #release-engineering-team, mmodell

Reviewed By: #release-engineering-team, mmodell

Subscribers: Dzahn, jenkins

Differential Revision: https://phabricator.wikimedia.org/D84