Move config_deploy to post-fetch
Summary:
Moves config_deploy stage to after fetch stage. Initially, since we
wanted a hash of the contents of the deployed revision, we had to run
config_deploy before any additional steps.

Now config_deploy relies soly on the deployed SHA1 changing. It makes
no attempt to rebuild any config if the SHA1 of the repo hasn't changed.
To update the config without updating the deployed SHA1 of the repo, use
the --force option.

Test Plan:
Changes to config that are not a new repo revision should no
longer deploy without the --force flag

Reviewers: dduvall, demon, mobrovac, #release-engineering-team, mmodell

Reviewed By: demon, mobrovac, #release-engineering-team, mmodell

Subscribers: jenkins, mobrovac

Differential Revision: https://phabricator.wikimedia.org/D88