Load extension-list from the branch if present.
Projects: #release-engineering-team

Maniphest Tasks: T125678

Summary:
This allows us to override the extension list on a per-branch basis,
which is useful when deploying or retiring an extension.

Previously we had to use dirty hacks to deploy when the extensions
differ between the two current branches. This is the first step to
solving that problem.

Step 2 is writing the extension-list when the branch is created.

refs T125678

Test Plan: do it live?

Reviewers: #release-engineering-team, demon, thcipriani, bd808

Reviewed By: bd808

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D115