Compare cdb mtimes less granularly
Projects: #release-engineering-team

Maniphest Tasks: T125108

Summary:
When the mtimes of the cdb files are set after they are rebuilt from the
json files via os.utime, there appears to be some strange truncating of
precision.

As a result, when these times are retrieved via os.path.getmtime the
exact comparison is a little off.

This patch moves to using a string comparison of the times.

Fixes T125108

Reviewers: #release-engineering-team, bd808, ori

Reviewed By: bd808, ori

Subscribers: mmodell, jenkins

Differential Revision: https://phabricator.wikimedia.org/D132