Use root context directory
Projects: #release-engineering-team

Summary:
If deploy-local is run like:

    sudo -u [deploy-user] -- deploy-local --repo [repo]

it will fail unless the deploy-user is able to cd into the current
directory. This is because the utils.cd context manager tries to
move back to the directory from which it started after the
cd context is closed.

This works fine for the case where a remote deploy user sshs to a target
box and runs a series of commands, but for the puppet provider, and for
people doing ad-hoc troubleshooting this is unworkable. This patch
should fix the bad observed behavior.

Test Plan:
* Try deploy-local as a user other than deploy_user using sudo
* Try deploy-local from a directory you have access to but the deploy-
  user does not.

Reviewers: demon, Ladsgroup, mmodell, #release-engineering-team

Reviewed By: mmodell, #release-engineering-team

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D208