Add git_rev config var
Projects: #release-engineering-team

Maniphest Tasks: T133572

Summary:
Allow user to specify revision to deploy via config file. This will
allow users to specify a particular ref to deploy in a particular
environment. For instance, it should now be possible to have a beta
branch that is deployed in the beta environment.

Fixes T133572

Reviewers: demon, hashar, mobrovac, #release-engineering-team, mmodell

Reviewed By: #release-engineering-team, mmodell

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D218