sync-wikiversions: Add missing sync steps
Tags: #release-engineering-team

Summary:
sync-wikiversions extends AbstractSync, but actually clobbers all of the
logic in AbstractSync::main. While looking at the code I vaguely
remember that this was done to stay more aligned with the prior
sync-wikiversions logic which skips using fanout rsync servers and
instead has the MediaWiki servers fetch directly from the deploy master.

A downside of this decision however is that more recent enhancements to
AbstractSync are not supported in sync-wikiversions. Most glaringly, the
co-master sync operation that keeps multiple deploy servers coordinated
is not part of sync-wikiversions before this patch.

Add support for:
* lock file protection around the sync operation
* respect for global sync flag
* co-master sync

Canary server deployment and logging checks from AbstractSync are still
not implemented at this point.

Test Plan: run sync-wikiverisons and see the co-master updated

Reviewers: Reedy, mmodell, thcipriani, demon, #release-engineering-team

Reviewed By: demon, #release-engineering-team

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D309