Support skipping remaining continue prompts
Tags: #release-engineering-team

Summary:
Follows up on T142990 to allow the user to continue through all
remaining deployment groups. This feature is necessary as there may be
many prompts if the repo is configured with a relatively small
`group_size`.

Test Plan:
Deploy using a number of server groups and ensure that all given
choices work as expected.

Reviewers: demon, mobrovac, mmodell, #release-engineering-team, thcipriani

Reviewed By: mmodell, #release-engineering-team, thcipriani

Subscribers: jenkins

Differential Revision: https://phabricator.wikimedia.org/D323