Make directories as euid
Tags: #release-engineering-team

Summary:
By default, mkdir_p was using the ruid of the user to create needed
directories. This causes breakage when a user locally calls something
like:

    sudo -u deploy-service deploy-local --repo test/repo config_deploy

Since, rather than creating the needed `[repo]/.git/config-deploy`
directory as the `ssh_user` from the config OR as `deploy-service` which
is the euid it uses the user that first logged into the system.

This doesn't cause any problems from the deployment-host, but will cause
problems if this command is run by a local user or by puppet.

The only place we aren't passing an explicit username from
the context object is in config_deploy.

Reviewers: mmodell, #release-engineering-team, demon, dduvall, mobrovac

Reviewed By: mobrovac

Differential Revision: https://phabricator.wikimedia.org/D343